requirements.txt
asgiref==3.6.0
attrs==22.2.0
autopep8==2.0.2
certifi==2022.12.7
cffi==1.15.1
chardet==5.1.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==39.0.2
Django==4.1.7
django-cors-headers==3.14.0
django-pytest==0.2.0
django-secure==1.0.2
django-sslserver==0.22
django-tabulate==0.1.1
django-test-without-migrations==0.6
djangorestframework==3.14.0
gunicorn==20.1.0
idna==3.4
iniconfig==2.0.0
itypes==1.2.0
Jinja2==3.1.2
MarkupSafe==2.1.2
mysql-connector-python==8.0.32
mysqlclient==2.1.1
netaddr==0.8.0
numpy==1.24.2
openapi-codec==1.3.2
packaging==23.0
pandas==1.5.3
pdfminer.six==20221105
pdfplumber==0.8.0
phonenumbers==8.13.7
Pillow==9.4.0
pluggy==1.0.0
protobuf==3.20.3
py==1.11.0
pycodestyle==2.10.0
pycparser==2.21
pycryptodome==3.17
pyparsing==3.0.9
pytest==7.2.2
python-dateutil==2.8.2
pytz<=2023.3
PyYAML==6.0
qrcode==7.4.2
requests==2.28.2
simplejson==3.18.4
six==1.16.0
sqlparse==0.4.3
tabulate==0.9.0
tblib==1.7.0
toml==0.10.2
tomli==2.0.1
uritemplate==4.1.1
urllib3==1.26.15
Wand==0.6.11
django-crontab==0.7.1
django-allow-cidr==0.6.0
reportlab==4.0.4
drf-yasg==1.21.7
djangorestframework-simplejwt==5.3.0
